Manutenção

+++ BibliotecaLPWSD.sql
@@ -12,14 +12,14 @@ SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,N
 CREATE SCHEMA IF NOT EXISTS `BibliotecaLPWSD` DEFAULT CHARACTER SET utf8 ;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Livro` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `titulo` VARCHAR(100) NOT NULL,
   `isbn` VARCHAR(9) NOT NULL,
   `edicao` INT(11) NOT NULL,
   `ano` INT(5) NOT NULL,
   `capa` VARCHAR(45) NULL DEFAULT NULL,
   `arquivo` VARCHAR(45) NULL DEFAULT NULL,
-  `idEditora` INT(11) NOT NULL,
+  `idEditora` BIGINT(11) NOT NULL,
   PRIMARY KEY (`id`),
   INDEX `fk_tbLivro_tbEditora1_idx` (`idEditora` ASC) ,
   UNIQUE INDEX `titulo_UNIQUE` (`titulo` ASC) ,
@@ -32,7 +32,7 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Assunto` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `assunto` VARCHAR(100) NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE INDEX `assunto_UNIQUE` (`assunto` ASC) )
@@ -40,9 +40,9 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Exemplar` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `circular` TINYINT(1) NULL DEFAULT NULL,
-  `idLivro` INT(11) NOT NULL,
+  `idLivro` BIGINT(11) NOT NULL,
   PRIMARY KEY (`id`),
   INDEX `fk_tbExemplar_tbLivro1_idx` (`idLivro` ASC) ,
   CONSTRAINT `fk_tbExemplar_tbLivro1`
@@ -54,12 +54,12 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Emprestimo` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `dataEmprestimo` DATE NOT NULL,
   `dataDevolucaoPrevista` DATE NOT NULL,
   `dataDevolucao` DATE NULL DEFAULT NULL,
-  `idExemplar` INT(11) NOT NULL,
-  `idUsuario` INT(11) NOT NULL,
+  `idExemplar` BIGINT(11) NOT NULL,
+  `idUsuario` BIGINT(11) NOT NULL,
   PRIMARY KEY (`id`),
   INDEX `fk_tbEmprestimo_tbExemplar1_idx` (`idExemplar` ASC) ,
   INDEX `fk_tbEmprestimo_tbUsuario1_idx` (`idUsuario` ASC) ,
@@ -77,7 +77,7 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Usuario` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `nome` VARCHAR(100) NOT NULL,
   `tipo` CHAR(1) NOT NULL,
   `email` VARCHAR(100) NOT NULL,
@@ -89,7 +89,7 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Editora` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `nome` VARCHAR(100) NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE INDEX `nome_UNIQUE` (`nome` ASC) )
@@ -97,7 +97,7 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Autor` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `nome` VARCHAR(100) NOT NULL,
   PRIMARY KEY (`id`) ,
   UNIQUE INDEX `nome_UNIQUE` (`nome` ASC) )
@@ -105,8 +105,8 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`AutorLivro` (
-  `idLivro` INT(11) NOT NULL,
-  `idAutor` INT(11) NOT NULL,
+  `idLivro` BIGINT(11) NOT NULL,
+  `idAutor` BIGINT(11) NOT NULL,
   PRIMARY KEY (`idLivro`, `idAutor`),
   INDEX `fk_tbLivro_has_tbAutores_tbAutores1_idx` (`idAutor` ASC) ,
   INDEX `fk_tbLivro_has_tbAutores_tbLivro1_idx` (`idLivro` ASC) ,
@@ -124,8 +124,8 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`LivroAssunto` (
-  `idLivro` INT(11) NOT NULL,
-  `idAssunto` INT(11) NOT NULL,
+  `idLivro` BIGINT(11) NOT NULL,
+  `idAssunto` BIGINT(11) NOT NULL,
   PRIMARY KEY (`idLivro`, `idAssunto`),
   INDEX `fk_Livro_has_Assunto_Assunto1_idx` (`idAssunto` ASC) ,
   INDEX `fk_Livro_has_Assunto_Livro1_idx` (`idLivro` ASC) ,
@@ -143,14 +143,14 @@ ENGINE = InnoDB
 DEFAULT CHARACTER SET = utf8;
 
 CREATE TABLE IF NOT EXISTS `BibliotecaLPWSD`.`Reserva` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(11) NOT NULL AUTO_INCREMENT,
   `dataReserva` DATE NOT NULL,
   `dataDevolucaoPrevista` DATE NOT NULL,
   `cancelada` TINYINT(1) NULL DEFAULT NULL,
   `obsCancelamento` VARCHAR(45) NULL DEFAULT NULL,
-  `idExemplar` INT(11) NOT NULL,
-  `idUsuario` INT(11) NOT NULL,
-  `idEmprestimo` INT(11) NULL DEFAULT NULL,
+  `idExemplar` BIGINT(11) NOT NULL,
+  `idUsuario` BIGINT(11) NOT NULL,
+  `idEmprestimo` BIGINT(11) NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
   INDEX `fk_Reserva_Exemplar1_idx` (`idExemplar` ASC) ,
   INDEX `fk_Reserva_Usuario1_idx` (`idUsuario` ASC) ,
