+++ src/main/java/br/cesjf/bibliotecalpwsd/model/Editora.java
@@ -38,7 +38,7 @@ public class Editora implements Serializable, Comparable<Editora> {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @Column(name = "nome")
     private String nome;
@@ -48,20 +48,20 @@ public class Editora implements Serializable, Comparable<Editora> {
     public Editora() {
     }
 
-    public Editora(Integer id) {
+    public Editora(Long id) {
         this.id = id;
     }
 
-    public Editora(Integer id, String nome) {
+    public Editora(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
