+++ src/main/java/br/cesjf/bibliotecalpwsd/model/Emprestimo.java
@@ -42,7 +42,7 @@ public class Emprestimo implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @Column(name = "dataEmprestimo")
     @Temporal(TemporalType.DATE)
@@ -64,19 +64,19 @@ public class Emprestimo implements Serializable {
     public Emprestimo() {
     }
 
-    public Emprestimo(Integer id) {
+    public Emprestimo(Long id) {
         this.id = id;
     }
 
-    public Emprestimo(Integer id, Date dataEmprestimo) {
+    public Emprestimo(Long id, Date dataEmprestimo) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -167,7 +167,7 @@ public class Emprestimo implements Serializable {
 
     @Override
     public String toString() {
-        return  Integer.toString(id);
+        return  Long.toString(id);
     }
     
 }
