+++ src/main/java/br/cesjf/bibliotecalpwsd/model/Exemplar.java
@@ -40,7 +40,7 @@ public class Exemplar implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "circular")
     private Boolean circular;
     @OneToMany(mappedBy = "idExemplar")
@@ -54,15 +54,15 @@ public class Exemplar implements Serializable {
     public Exemplar() {
     }
 
-    public Exemplar(Integer id) {
+    public Exemplar(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
